refactor(text-to-image): moves "Config/Static/exports.ts" to "Config/Static/read.ts"
