+++ src/features/TextToImage/Config/Static/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './read.ts';
